Refactor:공간 상세 방문 목적에 따라 아이콘 구별 및 컴포넌트로 구별
