+++ app/(pages)/(submain)/place/[id]/page.tsx
@@ -1,13 +1,13 @@
 import Divider from "@common/components/ui/divider/Divider";
-import PlaceDetailInfo from "@feature/place/components/PlaceDetail/PlaceDetailInfo";
+import PlaceDetailInfo from "@feature/place/components/PlaceDetail/organisms/PlaceDetailInfo";
 import PlaceDetailKeywordEvaluation from "@feature/place/components/PlaceDetail/PlaceDetailKeywordEvaluation";
 import PlaceDetailKeywordSummary from "@feature/place/components/PlaceDetail/PlaceDetailKeywordSummary";
 import GetPlaceDetail from "@feature/place/queries/getPlaceDetail";
-import PlaceDetailTopBar from "@feature/place/components/PlaceDetail/PlaceDetailTopBar";
-import PlaceInfoCard from "@feature/place/components/PlaceInfoCard/PlaceInfoCard";
+import PlaceDetailTopBar from "@feature/place/components/PlaceDetail/organisms/PlaceDetailTopBar";
+import PlaceInfoCard from "@feature/place/components/PlaceInfo/molecules/PlaceInfoCard";
 import SliderLayout from "@common/components/layout/SliderLayout/SliderLayout";
 import CurationCardDark from "@feature/curation/components/CurationCardDark/CurationCardDark";
-import PlaceImageSlider from "@feature/place/components/PlaceDetail/PlaceImageSlider";
+import PlaceImageSlider from "@feature/place/components/PlaceDetail/organisms/PlaceImageSlider";
 
 //Page
 export default async function PlaceDetailPage({
