+++ app/feature/place/components/PlaceDetail/PlaceDetailKeywordSummary.tsx
@@ -1,7 +1,6 @@
-import HeartIcon from "@common/assets/icons/heart/HeartIcon";
-import UserDouble from "@common/assets/icons/user/UserDouble";
 import Chip from "@common/components/ui/buttons/Chip/Chip";
 import { PlaceDetailInfoProps } from "@feature/place/queries/dto/place-detail";
+import PlaceDetailPurposeBox from "./molecules/PlaceDetailPurposeBox";
 
 export default function PlaceDetailKeywordSummary({
   mainText,
@@ -23,20 +22,8 @@ export default function PlaceDetailKeywordSummary({
       <div className="bg-background-secondary-light mt-[1.6rem] p-[2rem]">
         <span className="body2-medium text-text-gray-6">방문 목적</span>
         <div className="flex justify-between mt-[0.8rem] mb-[1.6rem]">
-          <div className="w-[48%] flex flex-col items-center p-[1.2rem] bg-white rounded-[8px]">
-            <HeartIcon color="#F670C7" />
-            <div className="mt-[0.8rem]">
-              <span className="body1 text-primary-normal"># </span>
-              <span className="body1 text-black">{purpose[0]}</span>
-            </div>
-          </div>
-          <div className="w-[48%] flex flex-col items-center p-[1.2rem] bg-white rounded-[8px]">
-            <UserDouble color="#9B8AFB" />
-            <div className="mt-[0.8rem]">
-              <span className="body1 text-primary-normal"># </span>
-              <span className="body1 text-black">{purpose[1]}</span>
-            </div>
-          </div>
+          <PlaceDetailPurposeBox purpose={purpose[0]} />
+          <PlaceDetailPurposeBox purpose={purpose[1]} />
         </div>
         {interior && interior[0].length > 0 && (
           <div className="mb-[0.8rem]">
