+++ app/feature/place/components/PlaceDetail/molecules/PlaceImageSliderCount.tsx
