+++ app/feature/place/components/PlaceDetail/organisms/PlaceDetailInfoMore.tsx
