+++ app/feature/place/components/PlaceDetail/organisms/PlaceDetailTopBar.tsx
