+++ app/feature/place/components/PlaceInfo/atoms/PlaceInfoCardTopScrapIcon.tsx
@@ -8,6 +8,7 @@ interface PlaceInfoCardTopScrapIconProps {
   onClick?: (e: React.MouseEvent<SVGSVGElement, MouseEvent>) => void;
 }
 
+//Atom
 export default function PlaceInfoCardTopScrapIcon({
   isScraped,
   cardSize,
