+++ app/feature/place/components/PlaceInfo/molecules/PlaceTypeAndAddress.tsx
