+++ app/feature/search/components/SearchResult/SearchResult.tsx
@@ -11,7 +11,7 @@ import {
 import Divider from "@common/components/ui/divider/Divider";
 import FilterIcon from "@common/assets/icons/filter/filter-keyword.svg";
 const PlaceInfoCard = lazy(
-  () => import("@feature/place/components/PlaceInfoCard/PlaceInfoCard")
+  () => import("@feature/place/components/PlaceInfo/molecules/PlaceInfoCard")
 );
 import SearchBar from "@feature/search/components/SearchBar/SearchBar";
 import HomeSearchSkeleton from "@feature/search/components/HomeSearchSkeleton/HomeSearchSkeleton";
